feat: add reusable fd cache
